requirements.txt
Flask==1.1.1
python-telegram-bot==11.1.0
requests==2.22.0